done! - còn nhiều lỗi, sau này khắc phục

+++ Models/VnPay/PaymentInformationModel.cs
@@ -3,7 +3,7 @@ namespace websiteTUTHIEN.Models.VnPay
     public class PaymentInformationModel
     {
         public string OrderType { get; set; } // values:other
-        public double Amount { get; set; } // field nhap so tien
+        public decimal Amount { get; set; } // field nhap so tien
         public string OrderDescription { get; set; } // values:tu thien vao web tu thien
         public string Name { get; set; } // field nguoi dung tu nhap
     }
